Fix signed/unsigned comparison warning

+++ include/rmm/device_buffer.hpp
@@ -324,7 +324,8 @@ class device_buffer {
    */
   [[nodiscard]] std::int64_t ssize() const noexcept
   {
-    assert(size() < std::numeric_limits<int64_t>::max() && "Size overflows signed integer");
+    assert(size() < static_cast<std::size_t>(std::numeric_limits<int64_t>::max()) &&
+           "Size overflows signed integer");
     return static_cast<int64_t>(size());
   }
